wip: rename app script
